gamma_budget/requirements.txt
django-jazzmin==3.0.0,<3.1
Django==5.0,<5.1
djangorestframework==3.15.1,<3.16
django-oauth-toolkit==2.3.0,<2.4
django-filter==24.2,<24.3
dj-database-url==2.2.0
psycopg[binary,pool]==3.1.18,<3.2
whitenoise==6.6.0,<6.7
pytest-django==4.8.0,<4.9
django-cors-headers==4.3.1,<4.4
